Move StreakComputer to a subdirectory
