+++ core/domain/src/main/java/com/rendox/routinetracker/core/domain/di/StreaksDomainModule.kt
@@ -2,8 +2,8 @@ package com.rendox.routinetracker.core.domain.di
 
 import com.rendox.routinetracker.core.domain.streak.GetAllStreaksUseCase
 import com.rendox.routinetracker.core.domain.streak.GetAllStreaksWithCashingUseCase
-import com.rendox.routinetracker.core.domain.streak.StreakComputer
-import com.rendox.routinetracker.core.domain.streak.StreakComputerImpl
+import com.rendox.routinetracker.core.domain.streak.computer.StreakComputer
+import com.rendox.routinetracker.core.domain.streak.computer.StreakComputerImpl
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
