+++ core/domain/src/main/java/com/rendox/routinetracker/core/domain/streak/computer/StreakComputerImpl.kt
@@ -1,4 +1,4 @@
-package com.rendox.routinetracker.core.domain.streak
+package com.rendox.routinetracker.core.domain.streak.computer
 
 import com.rendox.routinetracker.core.domain.habitstatus.HabitStatusComputer
 import com.rendox.routinetracker.core.domain.schedule.getPeriodRange
